refactor: :recycle: make attachments component more reusable
refactor the attachments component to be easyer to use in the extension

+++ src/components/BugTab/index.vue
@@ -1,6 +1,6 @@
 <template>
 	<div>
-		<SideTab v-if="bug" class="tab-shaddow">
+		<div v-if="bug" class="bs-tab bs-scroll s-purple" gap-4>
 			<Info
 				:bug="bug"
 				:status="status"
@@ -8,40 +8,51 @@
 				@open_assign="assignShow = true"
 			/>
 
-			<Attachments
-				:attachments="bug.attachments ? bug.attachments : []"
-				:bug_id="id"
-			/>
+			<AttachmentsList
+				:list="bug.attachments ?? []"
+				:error="attachments.error"
+				@update="attachments.update"
+				@upload="attachments.upload"
+			>
+				<template #item="{ item }">
+					<AttachmentsItem
+						:name="item.attributes.designation"
+						:id="item.id"
+						@download="attachments.download"
+						@delete="attachments.delete"
+					/>
+				</template>
+			</AttachmentsList>
 
 			<Comments
 				:comments="bug.comments ? bug.comments : []"
 				:bug_id="id"
 			/>
 
-			<div class="delete-btn-wrapper">
-				<div class="delete-bug-btn" @click="deleteBug">
-					<div class="delete-icon" />
+			<div class="delete-bug-btn" @click="deleteBug">
+				<img
+					src="/src/assets/icons/delete.svg"
+					alt="delete"
+					mr-2
+					class="black-to-red"
+				/>
 
-					<span class="text-capitalize">
-						{{ $t("delete.bug") }}
-					</span>
-				</div>
+				<span>
+					{{ $t("delete.bug") }}
+				</span>
 			</div>
-		</SideTab>
+		</div>
 
-		<AssignModal
-			v-if="assignShow"
-			:id="bug.id"
-			@close="assignShow = false"
-		/>
+		<AssignModal v-if="assignShow" :id="id" @close="assignShow = false" />
 
-		<div class="outside-overlay" @click="emit('close')" />
+		<div class="full-overlay" @click="emit('close')" />
 	</div>
 </template>
 
-<script setup>
+<script setup lang="ts">
 import { useProjectStore } from "~/stores/project";
 import axios from "axios";
+import toBase64 from "~/util/toBase64";
 
 const emit = defineEmits(["close", "deleted"]);
 const props = defineProps({
@@ -74,12 +85,14 @@ watch(bug, () => {
 });
 
 const deleteBug = async () => {
+	if (!status.value || !bug.value) return console.log("Strange error!");
+
 	try {
 		await axios.delete(`statuses/${status.value.id}/bugs/${bug.value.id}`);
 
-		status.value.attributes.bugs.splice(
+		status.value.attributes.bugs?.splice(
 			status.value.attributes.bugs.findIndex(
-				(x) => x.id === bug.value.id
+				(x) => x.id === bug.value?.id
 			),
 			1
 		);
@@ -91,54 +104,85 @@ const deleteBug = async () => {
 };
 
 const assignShow = ref(false);
+
+const attachments = reactive({
+	error: "",
+
+	upload: (files: File[]) => {
+		if (files.length > 0) {
+			files.forEach(async (file) => {
+				try {
+					let base64 = btoa(await toBase64(file));
+
+					axios
+						.post(`bugs/${props.id}/attachments`, {
+							designation: file.name,
+							base64: base64,
+						})
+						.then(() => {
+							attachments.update();
+						});
+				} catch (error: Error) {
+					attachments.error = error;
+					console.log(error);
+				}
+			});
+		}
+	},
+
+	download: (id: string) => {
+		axios
+			.get(`bugs/${props.id}/attachments/${id}`, {
+				headers: {
+					"include-attachment-base64": "true",
+				},
+			})
+			.then((response) => {
+				const link = document.createElement("a");
+				link.href = atob(response.data.data.attributes.base64);
+				link.setAttribute(
+					"download",
+					response.data.data.attributes.designation
+				); //or any other extension
+				link.id = "downloadAttachmentA";
+				document.body.appendChild(link);
+				link.click();
+				link.remove();
+			})
+			.catch((error) => {
+				console.log(error);
+			});
+	},
+
+	delete: (id: string) => {
+		axios
+			.delete(`bugs/${props.id}/attachments/${id}`)
+			.then(() => {
+				attachments.update();
+			})
+			.catch((error) => {
+				console.log(error);
+			});
+	},
+
+	update: () => {
+		store.fetchAttachments(props.id);
+	},
+});
 </script>
 
 <style lang="scss" scoped>
-.delete-btn-wrapper {
+.delete-bug-btn {
 	display: inline-flex;
 	align-items: center;
 	justify-content: center;
+	cursor: pointer;
 
-	.delete-bug-btn {
-		display: inline-flex;
-		align-items: center;
-		justify-content: center;
-		cursor: pointer;
-
-		> .delete-icon {
-			width: 24px;
-			height: 24px;
-			margin-right: 6px;
-			background-image: url("/src/assets/icons/delete.svg");
-			background-repeat: no-repeat;
-			background-position: center;
-
-			// color: #F23838; red
-			filter: brightness(0) saturate(1) invert(46%) sepia(28%)
-				saturate(5216%) hue-rotate(331deg) brightness(87%)
-				contrast(121%);
-		}
-
-		> span {
-			color: hsl(0, 90%, 60%);
-			text-transform: capitalize;
-			font-weight: 600;
-			font-size: 20px;
-		}
+	> span {
+		color: hsl(0, 90%, 60%);
+		text-transform: capitalize;
+		font-weight: 600;
+		font-size: 1.25rem;
 	}
-}
-
-.tab-shaddow {
-	box-shadow: -10px 0px 24px hsla(231, 42%, 18%, 0.11);
-}
-
-.outside-overlay {
-	position: fixed;
-	width: 100vw;
-	height: 100vh;
-	background: #00000073;
-	top: 0;
-	left: 0;
-	z-index: 1;
 }
 </style>
